layout/style/test/test_medialist_privilege.html

<!doctype html>
<title>Test whether it can access a filed in MediaList with normal privilege after accessing with chrome privilege.</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>

<style>
@media screen and (max-width: 800px) {
  body {
    background-color: lime;
  }
}
</style>
<script>
"use strict";

const chromeDocument = SpecialPowers.wrap(document);
assertMediaText(chromeDocument);
assertMediaText(document);

function assertMediaText(doc) {
  const stylesheet = doc.styleSheets[0];
  const rule = stylesheet.cssRules[0];
  const media = rule.media;
  is(rule.media.mediaText, "screen and (max-width: 800px)");
}
</script>
